+++ crates/solc-json-interface/src/standard_json/input/settings/mod.rs
@@ -1,6 +1,7 @@
 //! The `solc --standard-json` input settings.
 
 pub mod metadata;
+pub mod metadata_hash;
 pub mod optimizer;
 pub mod selection;