+++ crates/solc-json-interface/src/standard_json/mod.rs
@@ -1,0 +1,4 @@
+//! The `solc <input>.sol --standard-json` interface input and output.
+
+pub mod input;
+pub mod output;