+++ crates/solidity/src/solc/standard_json/mod.rs
@@ -1,4 +1,0 @@
-//! The `solc <input>.sol --standard-json`.
-
-pub mod input;
-pub mod output;